+++ grai-server/app/pyproject.toml
@@ -67,9 +67,10 @@ channels = "^4.0.0"
 channels-redis = "^4.1.0"
 uvicorn = {version = "^0.23.2", extras = ["standard"]}
 gunicorn = "^21.2.0"
-strawberry-graphql = {extras = ["channels"], version = "^0.209.7"}
+strawberry-graphql = {extras = ["channels", "asgi"], version = "^0.209.7"}
 asgi-cors-strawberry = "^0.2.0"
 overrides = "^7.4.0"
+wsproto = "^1.2.0"
 
 [tool.poetry.group.dev.dependencies]
 isort = "^5.10.1"
